Revert "Revert "Resource""
This reverts commit eaf9ed635d79c07b739009b36ee868052b3dcdfe.
