+++ com.sap.cloud.lm.sl.mta/src/main/java/com/sap/cloud/lm/sl/mta/model/v2/Resource.java
@@ -8,12 +8,27 @@ import java.util.Map;
 import org.apache.commons.lang3.ObjectUtils;
 
 import com.sap.cloud.lm.sl.common.util.MapUtil;
+import com.sap.cloud.lm.sl.mta.model.ElementContext;
+import com.sap.cloud.lm.sl.mta.model.NamedElement;
 import com.sap.cloud.lm.sl.mta.model.ParametersContainer;
+import com.sap.cloud.lm.sl.mta.model.PropertiesContainer;
+import com.sap.cloud.lm.sl.mta.model.VisitableElement;
+import com.sap.cloud.lm.sl.mta.model.Visitor;
 import com.sap.cloud.lm.sl.mta.parsers.v2.ResourceParser;
 import com.sap.cloud.lm.sl.mta.util.YamlElement;
 
-public class Resource extends com.sap.cloud.lm.sl.mta.model.v1.Resource implements ParametersContainer {
+public class Resource implements VisitableElement, NamedElement, PropertiesContainer, ParametersContainer {
 
+    @YamlElement(ResourceParser.NAME)
+    private String name;
+    @YamlElement(ResourceParser.TYPE)
+    private String type;
+    @YamlElement(ResourceParser.DESCRIPTION)
+    private String description;
+    @YamlElement(ResourceParser.GROUPS)
+    private List<String> groups;
+    @YamlElement(ResourceParser.PROPERTIES)
+    private Map<String, Object> properties;
     @YamlElement(ResourceParser.PARAMETERS)
     private Map<String, Object> parameters;
 
@@ -21,27 +36,62 @@ public class Resource extends com.sap.cloud.lm.sl.mta.model.v1.Resource implemen
 
     }
 
-    @Override
+    public String getName() {
+        return name;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+ 
     public List<String> getGroups() {
         throw new UnsupportedOperationException();
     }
 
     @Override
+    public Map<String, Object> getProperties() {
+        return MapUtil.unmodifiable(properties);
+    }
+    
+    @Override
     public Map<String, Object> getParameters() {
         return MapUtil.unmodifiable(parameters);
     }
+    
+    public void setName(String name) {
+        this.name = name;
+    }
 
-    @Override
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
     public void setGroups(List<String> groups) {
         throw new UnsupportedOperationException();
+    }
+    
+    public void setProperties(Map<String, Object> properties) {
+        this.properties = new LinkedHashMap<>(properties);
     }
 
     @Override
     public void setParameters(Map<String, Object> parameters) {
         this.parameters = new LinkedHashMap<>(parameters);
     }
-
+    
     @Override
+    public void accept(ElementContext context, Visitor visitor) {
+        visitor.visit(context, this);
+    }
+
     public Resource copyOf() {
         Builder result = new Builder();
         result.setName(getName());
@@ -52,11 +102,15 @@ public class Resource extends com.sap.cloud.lm.sl.mta.model.v1.Resource implemen
         return result.build();
     }
 
-    public static class Builder extends com.sap.cloud.lm.sl.mta.model.v1.Resource.Builder {
+    public static class Builder {
 
+        protected String name;
+        protected String type;
+        protected String description;
+        protected List<String> groups;
+        protected Map<String, Object> properties;
         protected Map<String, Object> parameters;
 
-        @Override
         public Resource build() {
             Resource result = new Resource();
             result.setName(name);
@@ -67,15 +121,30 @@ public class Resource extends com.sap.cloud.lm.sl.mta.model.v1.Resource implemen
             return result;
         }
 
-        public void setParameters(Map<String, Object> parameters) {
-            this.parameters = parameters;
+        public void setName(String name) {
+            this.name = name;
         }
 
-        @Override
+        public void setType(String type) {
+            this.type = type;
+        }
+
+        public void setDescription(String description) {
+            this.description = description;
+        }
+
         public void setGroups(List<String> groups) {
             throw new UnsupportedOperationException();
         }
+        
+        public void setProperties(Map<String, Object> properties) {
+            this.properties = properties;
+        }
 
+        public void setParameters(Map<String, Object> parameters) {
+            this.parameters = parameters;
+        }
+        
     }
 
 }
